+++ src/components/badge/badge.js
@@ -1,6 +1,6 @@
 import mergeData from 'vue-functional-data-merge'
 
-import { pluckProps } from '../utils/pluckProps'
+import pluckProps from '../../utils/pluck-props'
 import { assign } from '../../utils/object'
 import Link, { propsFactory as linkPropsFactory } from '../link/link'
