+++ src/components/breadcrumb/breadcrumb-link.js
@@ -1,5 +1,5 @@
 import mergeData from 'vue-functional-data-merge'
-import { pluckProps } from '../utils/pluckProps'
+import { pluckProps } from '../utils/pluck-props'
 import { assign } from '../../utils/object'
 import Link, { propsFactory as linkPropsFactory } from '../link/link'
